Move BestStatementsFinder into Statement namespace
